+++ packages/components/ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `Link`: fixed behaviour of `style` when `theme.color` is nullish
 - add `MultiBanner` component
+- add `NoScript` component
 
 ## `@svizzle/ui` v0.4.0