+++ packages/components/ui/CHANGELOG.md
@@ -16,7 +16,9 @@
 - add `FontsLoader` component
 - replaced some absolute CSS definitions related to text and fonts
 - added missing components to `rollup` input variable
-- fixed linting errors in `src/a11y/menu/settings.js`
+- fixed linting errors:
+	- in `src/a11y/menu/settings.js`
+	- linting now fails on warnings
 
 ## `@svizzle/ui` v0.4.0